re-organize data file structure
