+++ portal/static/js/src/components/PsaTracker.vue
@@ -137,7 +137,7 @@
     import tnthAjax from "../modules/TnthAjax.js";
     import tnthDates from "../modules/TnthDate.js";
     import SYSTEM_IDENTIFIER_ENUM from "../modules/SYSTEM_IDENTIFIER_ENUM.js";
-    import PsaTrackerData from "../data/PsaTrackerData.js";
+    import PsaTrackerData from "../data/gil/PsaTrackerData.js";
     export default {
         components: { RefreshIcon, FilterControl},
         errorCaptured: function(Error, Component, info) { /*eslint no-console: off */
