+++ portal/static/js/src/data/common/AssessmentReportData.js
