+++ portal/static/js/src/data/common/ExportInstrumentsData.js
